Collapse sharding v3 to plan/filter/merge + one research driver.
Keep history×cpu LPT in plan.py, graph slice in filter.py, report merge in
merge.py; drop suite-list/shell/nav helpers and the old parallel/debug
workflow names. Wire PR-check sharding to the slim scripts and simplify
test_ya in-shard retries to the main graph path.
