+++ .github/actions/test_ya/action.yml
@@ -542,48 +542,17 @@ runs:
           CURRENT_REPORT=$CURRENT_PUBLIC_DIR/report.json
 
           RETRY_YA_MAKE_TARGET="$YA_MAKE_TARGET"
-          RERUN_EXTRA_OPTS=()
+          RERUN_FAILED_OPT=""
           if [ $IS_RETRY = 0 ]; then
             CURRENT_MESSAGE="$CURRENT_MESSAGE"
           else
             CURRENT_MESSAGE="$CURRENT_MESSAGE (failed tests rerun, try $RETRY)"
-            MUTED_YAML_PATH="$CURRENT_PUBLIC_DIR/muted_tests.yaml"
-            python3 .github/scripts/tests/mute/mute_utils.py convert_muted_txt_to_yaml "$MUTED_YA_FILE" "$PREV_REPORT" > $MUTED_YAML_PATH
-            RERUN_EXTRA_OPTS+=(--build-only-test-deps)
-            if [ "$CROSS_JOB_RETRY" = "1" ] && [ "$RETRY" = "$START_RETRY" ]; then
-              # Build-only saved graph ignores --test-filter; run explicit suite targets instead.
-              GRAPH_PATH=""
-              CONTEXT_PATH=""
-              RETRY_YA_MAKE_TARGET=""
-              while IFS= read -r suite; do
-                if [ -n "$suite" ] && [ -e "$suite" ]; then
-                  RETRY_YA_MAKE_TARGET="$RETRY_YA_MAKE_TARGET $suite"
-                fi
-              done < <(python3 .github/scripts/tests/sharding/extract_failed_suite_paths.py "$PREV_REPORT")
-              if [ -z "$(echo "$RETRY_YA_MAKE_TARGET" | tr -d ' ')" ]; then
-                echo "Cross-job retry: no suite paths resolved, falling back to original build_target"
-                RETRY_YA_MAKE_TARGET="$YA_MAKE_TARGET"
-              else
-                echo "Cross-job retry: suite targets:$RETRY_YA_MAKE_TARGET"
-              fi
-            fi
             if [ -z "$GRAPH_PATH" ]; then
-              RERUN_EXTRA_OPTS+=(--test-blacklist-path "$MUTED_YAML_PATH")
-            fi
-            if [ "$CROSS_JOB_RETRY" = "1" ] && [ "$RETRY" = "$START_RETRY" ]; then
-              FAILED_FILTER_COUNT=0
-              while IFS= read -r filter; do
-                if [ -n "$filter" ]; then
-                  RERUN_EXTRA_OPTS+=(--test-filter "$filter")
-                  FAILED_FILTER_COUNT=$((FAILED_FILTER_COUNT + 1))
-                fi
-              done < <(python3 .github/scripts/tests/sharding/extract_failed_test_filters.py "$PREV_REPORT")
-              echo "Cross-job retry: scheduling $FAILED_FILTER_COUNT failed test(s) via --test-filter"
-              if [ "$FAILED_FILTER_COUNT" = "0" ]; then
-                RERUN_EXTRA_OPTS+=(-X)
-              fi
+              MUTED_YAML_PATH="$CURRENT_PUBLIC_DIR/muted_tests.yaml"
+              python3 .github/scripts/tests/mute/mute_utils.py convert_muted_txt_to_yaml "$MUTED_YA_FILE" "$PREV_REPORT" > $MUTED_YAML_PATH
+              RERUN_FAILED_OPT="-X --build-only-test-deps --test-blacklist-path $MUTED_YAML_PATH"
             else
-              RERUN_EXTRA_OPTS+=(-X)
+              RERUN_FAILED_OPT=""
             fi
           fi
 
@@ -620,7 +589,7 @@ runs:
 
           YA_MAKE_CMD=(
             ./ya make "${params[@]}" $RETRY_YA_MAKE_TARGET $GRAPH_OPTS
-            "${RERUN_EXTRA_OPTS[@]}" --log-file "$PUBLIC_DIR/ya_log.txt"
+            $RERUN_FAILED_OPT --log-file "$PUBLIC_DIR/ya_log.txt"
             --evlog-file "$CURRENT_PUBLIC_DIR/ya_evlog.jsonl"
             --build-results-report "$CURRENT_REPORT" --output "$YA_MAKE_OUT_DIR"
           )
