+++ .github/workflows/pr_check_sharding.yml
@@ -1,10 +1,6 @@
-name: Run and debug tests
-# Driver for sharding v3 research pipeline (pr_check_parallel_core.yml).
-#
-# Runs the pipeline over a real PR (checkout of its merge commit, like
-# PR-check does) or over a branch, without publishing anything to the PR,
-# and prints a duration summary to compare against the production PR-check
-# run of the same PR.
+name: PR-check sharding
+# Sharding v3 research driver: resolve PR/branch → plan/filter/merge pipeline.
+# Uses pr_check_sharding_core.yml (main test_ya + plan.py / filter.py / merge.py).
 
 on:
   workflow_dispatch:
@@ -136,8 +132,8 @@ jobs:
 
   run_check:
     needs: resolve
-    name: Parallel check
-    uses: ./.github/workflows/pr_check_parallel_core.yml
+    name: Sharded check
+    uses: ./.github/workflows/pr_check_sharding_core.yml
     secrets: inherit
     with:
       checkout_ref: ${{ needs.resolve.outputs.checkout_ref }}
